refactor : useAutoAdjustStage 커스텀 훅을 통한 마인드맵 위치 재조정 로직 분리
